Introduce BlockReaderCache and TableReaderCache
As their names suggest, BlockReaderCache and TableReaderCache are used to cache BlockReader and TableReader.
With regard to TableReaderCache, it helps to avoid calling system call each time a TableReader is used to search key/value pairs.
About BlockReaderCache, when searching a key in block in a SST file, if this block hasn't been in cach yet, BlockReaderCache will add it into cache. This prevents from recreating new BlockReader each time it is required.